Move `WasmOptHandler` into its own module
