+++ src/main.rs
@@ -20,6 +20,7 @@ mod cmd;
 mod crate_metadata;
 mod util;
 mod validate_wasm;
+mod wasm_opt;
 mod workspace;
 
 use self::{
